mercai-backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Environment
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication
PyJWT==2.8.0
bcrypt==4.1.1

# Validation
marshmallow==3.20.1

# HTTP Requests
requests==2.31.0

# Web Scraping
beautifulsoup4==4.12.2

# Cache
redis==5.0.1

# AI
google-generativeai==0.3.1

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
